chore: Rename page components
