+++ app/javascript/dashboard/routes/dashboard/campaigns/pages/SMSCampaignsPage.vue
